Pull wave dependencies into project source
This will allow us to delete the nasty jars in the repo.
The actual footprint from wave is relatively small,
so by pulling in the source, we not only don't have to
keep big binary files in gwt, we also don't have to parse
a ton of unused code at compile time, and we can make
updates to code that was otherwise unavailable.