Code cleanup and rebase on latest xapi
This ugly commit has a bunch of accumulated tweaks all over the project.

Of note, the original prototype of xapi server was built here,
then moved into xapi proper; so, some artifacts of that prototype still
live on, and we are committing them for posterity.
There will be a cleanup of unused artifacts later.